Move the get_labeled_nodes function to the machien.py
